refactor(PlayerName): 플레이어 이름을 저장하는 클래스의 패키지 이동
- /model/ 패키지에서 /model/role 패키지로 이동한다

- Side Effect
기존에 존재했던 클래스와 새로 추가한 클래스에서 PlayerName 참조 위취가 변경된다
