+++ src/main/kotlin/blackjack/model/Player.kt
@@ -2,6 +2,7 @@ package blackjack.model
 
 import blackjack.model.card.CardHand
 import blackjack.model.card.CardHandState
+import blackjack.model.role.PlayerName
 
 data class Player(override val name: PlayerName, override val cardHand: CardHand) : Role(name, cardHand) {
     override fun getState(): CardHandState {
