+++ src/main/kotlin/blackjack/model/PlayerWinning.kt
@@ -1,5 +1,7 @@
 package blackjack.model
 
+import blackjack.model.role.PlayerName
+
 data class PlayerWinning(val result: Map<PlayerName, WinningResultStatus>) {
     fun judgeDealerWinningResult(): DealerWinning =
         DealerWinning(
