+++ src/main/kotlin/blackjack/model/Role.kt
@@ -2,6 +2,7 @@ package blackjack.model
 
 import blackjack.model.card.CardHand
 import blackjack.model.card.CardHandState
+import blackjack.model.role.PlayerName
 
 abstract class Role(open val name: PlayerName, open val cardHand: CardHand) {
     abstract fun getState(): CardHandState
