+++ src/main/kotlin/blackjack/model/role/Dealer.kt
@@ -1,7 +1,5 @@
 package blackjack.model.role
 
-import blackjack.model.PlayerName
-
 class Dealer : Role(name = PlayerName(DEALER)) {
     val dealerDecisionCondition = { state.getCardHands().calculateScore() <= DEALER_MAX_SCORE_FOR_HIT }
