+++ src/main/kotlin/blackjack/model/role/Player.kt
@@ -1,5 +1,3 @@
 package blackjack.model.role
 
-import blackjack.model.PlayerName
-
 class Player(override val name: PlayerName) : Role(name)
