+++ src/test/kotlin/blackjack/model/ParticipantsTest.kt
@@ -4,6 +4,7 @@ import blackjack.model.card.Card
 import blackjack.model.card.CardHand
 import blackjack.model.card.CardNumber
 import blackjack.model.card.CardShape
+import blackjack.model.role.PlayerName
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
